speed up cpp version - via @owbone
Amazing that there's a 10% speedup in both g++/clang++ by just adding or
deleting a default constructor. We're not even using that. Deleting it
kind of makes sense to do as a security thing though, so allowing this
as a good semantic addition.

This is the fastest one now normally, there seems to be more variance in
the performance of this so have put ~300ms. It's only implementation
that I've seen go below 300.