Add RPython implementation
What is RPython?
----------------

"RPython is a restricted subset of Python2 that is amenable to static
analysis" - rpython.readthedocs.io/en/latest/rpython.html

No, you really shouldn't be writing programs in it. It's really
designed for writing language interpreters.

It is distributed as part of the pypy_ project.

.. _pypy: http://pypy.org/

What is this?
-------------

An implementation of `forest.py`, adapted to the RPython toolchain

Changes needed:

  - Add `entry_point` (where the compiled binary starts)
  - Add `target` (where the rpython compiler starts)
  - Add compatibility `__name__ == __main__` (where python can start)
  - No use of the standard library, not even `sys`. Namedtuples have
    been replaced with tuples (this might be a rule breaker).
  - Reduced string formatting options, you basically have str/str
    concat, and C-style formatting (if you need to convert ints).
  - RPython has no implementation of `any` or `set`.
  - RPython does have generators, but they're not very complex.
  - Watch out for functions that return multiple types.
    e.g. avoid returning both lists and iterators, as they don't
    share a subclass (and thus are not allowed). It's probably
    more efficient in RPython to just use preallocated lists.

How do I run it?
----------------

RPython programs are valid Python programs. You could just run
this under a normal python interpreter. Since the language is
simpler for an interpreter to reason with, you might get a boost
in performance anyway. The tradeoff is that an RPython implementation
is no longer an idiomatic Python implementation.
I'm not even sure if "idiomatic RPython" is a phraise that should exist.

How do I compile it?
--------------------

You will need a copy of the RPython toolchain. It is implemented
in python, so you don't need much more than to clone the pypy source.

    $ hg clone https://bitbucket.org/pypy/pypy ../pypy

The rpython toolchain in in the `rpython` subdirectory.

    $ ../pypy/rpython/bin/rpython ./forest_r.py

This will output `forest_r-c`.

    $ time ./forest_r-c 305 295 300