simpler-ish rpython installation
At least we don't need a user to compensate for AUR restrictions, and
we'll also get an up to date version now as aur pin of rpython is pretty
old anyway.

We install from source tarball at latest pypy2 version (5.8.0 atm).
Note that this needs `make` to run, and will just crash silently if you don't have
it.